requirements.txt
# Core dependencies
langchain-text-splitters>=0.2.1
sentence-transformers>=3.2.0
pydantic-ai>=1.9.1
numpy>=1.26.4
pandas>=2.2.2
spacy>=3.7.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
langchain-text-splitters>=0.1.0
google-generativeai>=0.4.0
presidio-analyzer>=2.2.355
presidio-anonymizer>=2.2.355
pydantic-ai
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-api
pydantic-evals
logfire

asyncio
json5
typing-extensions
uuid
glob2
os-sys

